+++ FluentSecurity/SecurityConfiguration.cs
@@ -18,7 +18,7 @@ namespace FluentSecurity
 			Advanced = Expression.Advanced;
 			ExternalServiceLocator = Expression.ExternalServiceLocator;
 			IgnoreMissingConfiguration = Expression.ShouldIgnoreMissingConfiguration;
-			PolicyContainers = Expression;
+			PolicyContainers = Expression.PolicyContainers;
 		}
 
 		internal ConfigurationExpression Expression { get; private set; }